aaaaaaaaa

+++ replot_all.py
@@ -3,23 +3,24 @@ import yaml
 from components.analysis import Analysis
 
 def delete_pdfs(dirname: str, namestring):
-    for filename in os.listdir(dirname):
-        if filename.endswith(".pdf") and namestring in filename:
-            os.remove(os.path.join(dirname, filename))
-            print(f"Deleted {dirname} {filename}")
+    if os.path.exists(dirname):
+        for filename in os.listdir(dirname):
+            if filename.endswith(".pdf") and namestring in filename:
+                os.remove(os.path.join(dirname, filename))
+                print(f"Deleted {dirname} {filename}")
 
 def delete_old_plots(dirname: str, which_plots):
     for plot in which_plots:
         if plot == "time":
-            delete_pdfs(dirname+"/records", "state")
+            delete_pdfs(dirname+"/records/time/", "")
         elif plot == "boxplots":
-            delete_pdfs(dirname+"/records", "box")
+            delete_pdfs(dirname+"/records/box/", "")
         elif plot == "losses":
             delete_pdfs(dirname+"/records/loss/", "main")
         elif plot == "gradients":
-            delete_pdfs(dirname+"/records/loss/", "gradient")
+            delete_pdfs(dirname+"/records/loss/", "gradient_")
         elif plot == "runs":
-            delete_pdfs(dirname+"/records/runs/", "run")
+            delete_pdfs(dirname+"/records/runs/", "")
         elif plot == "collisions":
             delete_pdfs(dirname+"/records/", "collisions")
 
@@ -35,11 +36,10 @@ plots = {
 analysis_configs: list[str] = [
     # -------- AICON vs. control --------
     # "exp1_sine.yaml",
-    # "exp1_extended.yaml",
+    
     # # ---------- Disturbances ----------
     # # Dist Sensor Failure
     # "exp2_distloss_sine.yaml",
-    "exp2_distloss_extended.yaml",
     # # Dist Noise
     # "exp2_non0mean_stat.yaml",
     # "exp2_non0mean_sine.yaml",
@@ -48,6 +48,11 @@ analysis_configs: list[str] = [
     # "exp2_wind_sine_nodist.yaml",
     # "exp2_wind_sine_distdot.yaml",
     # "exp2_wind_stat_distdot.yaml",
+
+    # "exp1_extended.yaml",
+    "exp2_distloss_extended.yaml",
+    # "exp2_non0mean_extended.yaml",
+    # "exp2_wind_extended.yaml",
 ]
 
 which_plots = [
@@ -69,6 +74,7 @@ if __name__ == "__main__":
         found_counter = 0
         for subfolder_name in os.listdir(records_path):
             if '_'.join(subfolder_name.split('_')[:-1]) == custom_name:
+                print("Replotting ", subfolder_name)
                 found_counter += 1
                 dirname = os.path.join(records_path, subfolder_name)
                 delete_old_plots(dirname, which_plots)
